requirements.txt
accelerate
bitsandbytes
comet-ml
datasets
evaluate
fast-langdetect
huggingface_hub
huggingface_hubtorch
langdetect
numpy
pandas
peft
sacrebleu
scikit-learn
sentencepiece
torch
tqdm
transformers
